Changed BLTouch Offsets
Note that Z height will have to be corrected whenever I change hot end - a tool for setting the BLtouch to a consistent height over bed will be handy!

Also increased x, y & z homing feedrates because they are a little slow for the X3 with its large endstop switches.